utils.py: hack around chardet
external package chardet was being used to guess encoding of strings so
that they can be decoded.  This is only done in utils.make_compat_str,
which is only called from TagExtractor.render_str().  It seems to me
that trying to guess an encoding for each substring has to be wrong (not
to mention expensive).  The encodng is most likely specified (or
asserted) at a much higher level.  So, while this change is a hack and
likely not correct for all PDFs, it's likely no worse than what was
there before.